# specs/001-web-automation-checker/contracts/api.yaml
openapi: 3.0.3
info:
  title: Web Automation Checker API
  description: |
    REST API for the Web Automation Checker tool. Allows users to submit URLs for automated testing,
    retrieve test reports, and view historical test data.

    **Features**:
    - P1: Basic URL testing with UI and performance checks
    - P2: Detailed diagnostic information for failures
    - P3: Test history and trend analysis
  version: 1.0.0
  contact:
    name: Anita Project Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://checker.example.com/api/v1
    description: Production server
...
paths:
  /tests:
    post:
      tags:
        - Tests
      summary: Create a new test request
      description: |
        Submit a URL for automated testing. The system will perform UI functionality checks
        and performance measurements, then return a test report.

        **User Story**: P1 - Basic URL Testing
      operationId: createTest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestRequest'
            examples:
              basic:
                summary: Basic URL test
                value:
                  url: "https://example.com"
              with_config:
                summary: Test with custom configuration
                value:
                  url: "https://example.com"
                  config:
                    timeout: 45
                    waitTime: 7
      responses:
        '201':
          description: Test request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRequest'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                url: "https://example.com"
                requestedAt: "2025-11-27T10:30:00Z"
                status: "pending"
                config:
                  timeout: 30
                  waitTime: 5
        '400':
          description: Invalid request (bad URL format, invalid config)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Validation Error"
                message: "Invalid URL format. Must be http:// or https://"
                details:
                  - field: "url"
                    issue: "Must match pattern ^https?://"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Rate Limit Exceeded"
                message: "Too many requests. Max 10 per minute."
                retryAfter: 45
  /tests/{testId}:
    get:
      tags:
        - Tests
      summary: Get test request status
      description: |
        Retrieve the current status of a test request. Poll this endpoint to monitor
        test execution progress.
      operationId: getTestStatus
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Test request ID
      responses:
        '200':
          description: Test request found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRequest'
        '404':
          description: Test request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    CreateTestRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          pattern: '^https?://'
          maxLength: 2048
          description: URL to test
          example: "https://example.com"
        config:
          type: object
          properties:
            timeout:
              type: integer
              minimum: 5
              maximum: 60
              default: 30
              description: Test timeout in seconds
            waitTime:
              type: integer
              minimum: 0
              maximum: 10
              default: 5
              description: Wait time for JavaScript execution (seconds)
    TestRequest:
      type: object
      required:
        - id
        - url
        - requestedAt
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique test request identifier
        url:
          type: string
          format: uri
          description: The URL being tested
        requestedAt:
          type: string
          format: date-time
          description: When the test was requested
        status:
          type: string
          enum: [pending, running, completed, failed]
          description: Current status of the test
        config:
          type: object
          properties:
            timeout:
              type: integer
            waitTime:
              type: integer
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
          description: Validation error details
        retryAfter:
          type: integer
          description: Seconds to wait before retrying (for rate limits)
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future enhancement)
security: []  # No authentication required for MVP